# docs/openAPI/openapi/payments.yaml
openapi: 3.0.1
info:
  title: EvenToNight - Ticketing API
  version: 1.0.0
  description: API for managing event ticket types, pricing, and availability
servers:
  - url: http://localhost:9050
    description: Local server
  - url: https://ticketing.eventonight.site
    description: Production server
paths:
...
  /orders/{orderId}:
    get:
      summary: Get order details
      description: Get details for a specific order (user must own the order)
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
          example: order_123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: order_123e4567-e89b-12d3-a456-426614174000
                  userId:
                    type: string
                    example: usr_123e4567-e89b-12d3-a456-426614174000
                  ticketIds:
                    type: array
                    items:
                      type: string
                    example: ["tick_123e4567-e89b-12d3-a456-426614174000"]
                  status:
                    type: string
                    enum: [PENDING, COMPLETED, CANCELLED]
                    example: COMPLETED
                  createdAt:
                    type: string
                    format: date-time
                    example: 2024-01-01T12:00:00Z
              example:
                id: order_123e4567-e89b-12d3-a456-426614174000
                userId: usr_123e4567-e89b-12d3-a456-426614174000
                ticketIds:
                  - tick_123e4567-e89b-12d3-a456-426614174000
                  - tick_456e7890-e89b-12d3-a456-426614174111
                status: COMPLETED
                createdAt: 2024-01-01T12:00:00Z
        '403':
          description: Forbidden - User is not authorized to access this order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: User is not authorized to access this order
                error: Forbidden
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: Order not found
                error: Not Found
      tags:
        - Orders
  /orders/{orderId}/pdf:
    get:
      summary: Get order PDF
      description: Download a PDF for the specified order (user must own the order)
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
          example: order_123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: Order PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '403':
          description: Forbidden - User is not authorized to access this order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: User is not authorized to access this order
                error: Forbidden
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: Order not found
                error: Not Found
      tags:
        - Orders
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Error message or array of validation errors
          example: "Validation failed"
        error:
          type: string
          description: Error type
          example: Bad Request
      required:
        - statusCode
        - message
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enter your JWT token in the format: Bearer <token>'